In the menu, replaced Products by Portail and Values by Webmail.
In the past, the https://calcool.ai/mail path was served by Apache2 configured as a reverse proxy.
The webmail is a PHP server running on Apache. /mail is an alias to /var/run/roundcube so we only need to pass the call to Apache.

The portal sits at http://openerp.lxd:8069/ and requires several paths to work: /portal/ converted to / /auth_signup/ /web/ /edi/ /board/ /web_linkedin/ /web_diagram/ /longpolling/.